fix for SCHEDULING-423
fix for SCHEDULING-579 on Windows 32 bits, On Windows 64 bits, the fix is still valid but needs an update of the winp.jar library as explained in the Bug description.

Improved reliability of find_matlab_command scripts

Improved debug logging 

Improved configure and makefile scripts for Matlab versions 7.7 to 7.9

Made ptmatlab.cc 7.7 to 7.9 compliant.  

Added ptolemy libraries for : 
- Matlab 7.7 : linux64 bits, Windows 32, Windows 64
- Matlab 7.9 : linux64 bits, Windows 64 bits


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@16437 28e8926c-6b08-0410-baaa-805c5e19b8d6