SCHEDULING-1472 RM and Scheduler GUI clients: Increase ping timeout and add busy indicator
For current release it was decided to not change Scheduler's and ResourceManager's methods used for pining to ImmediateServices. This means that pinging becomes less reliable and as a workaround ping timeout was increased to 30 seconds. Also now GUI isn't blocked when server call is in progress and there were no any indication that operation is executed. To improve user experience I decided to show busy cursor during the time when some server call is in progress.
I added in the ActiveObjectProxy support for events: if some call on active object is started the 'activityStarted' event is fired and when call is finished then event 'activityFinished' is fired. I implemented common listener for both GUI clients - ActivityIndicatorSupport which listens for these events and shows/hides busy cursor. ActivityIndicatorSupport can be enabled/disabled, this is needed when operation is in progress in one perspective and user switches to another one, in this case busy cursor should be hidden for the time when user works in the second perspective (it is disabled by existing PerspectiveAdapters when perspective changes)    


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@21368 28e8926c-6b08-0410-baaa-805c5e19b8d6