SCHEDULING-1476 : Add ClientJobState and ClientTaskState
2 classes are added, ClientJobState (extends JobState) and
ClientTaskState (extends TaskState).

They are instantiated in the scheduler core before being sent to the
frontend.

ClientJobState#update(TaskInfo) does the same as
InternalJob#update(TaskInfo).

ClientJobState#update(JobInfo) is similar to
InternalJob#update(JobInfo), but it differs in handling new tasks or
updates to existing task dependencies that might appear as a result of
control flow actions.

In InternalJob#update(JobInfo), the control flow actions are sort of
replayed one more time in the long updateTasksLooped() and
updateTasksReplicated() methods. Instead of doing that, I chose to
directly send to the frontend the new ClientTaskState instances that are
created or modified during control flow actions. Potentially this way we
could be sending more data between core and frontend which could hurt
performance, but it is hard to tell without measurements. At least it
simplifies the code a great deal.

To send new and modified ClientTaskStates in the update events, I've
added the JobInfoImpl#setModifiedTasks() and
JobInfoImpl#getModifiedTasks() methods.

The InternalJob#terminateTask() method is changed to keep track of the
new and modified tasks and to use jobInfo.setModifiedTasks() instead of
jobInfo.setTasksLooped() and jobInfo.setTasksReplicated().

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@21383 28e8926c-6b08-0410-baaa-805c5e19b8d6