BUILD-2 Remove unnecessary jars stored in svn, improve dependencies management in the ProActive projects
Removed content of the 'lib/ProActive' in the scheduling project:

Content of the folder 'scheduling/lib/ProActive' is produced as result of the ProActive Programming compilation, it is just a copy of the 'Programming/dist/lib', so it isn't necessary to store in the svn all these copied files. Scheduling project has special ant target 'copy.dependencies' which copies required jars from the Programming and after this change to compile scheduling it is necessary first execute 'copy.dependencies' target (or just copy 'Programming/dist/lib' into 'lib/ProActive' manually). 


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/trunk@21433 28e8926c-6b08-0410-baaa-805c5e19b8d6