SCHEDULING-1592 Improve error reporting for forked jvm launching
Problem:
To stark forked java task ForkedJavaExecutable creates several temporary
configuration files, these file are deleted when task completes, but if
it would be good to get content of these files in case if forked jvm
fails to start. 

Fix:
In this change I just modified ForkedJavaExecutable to include in the
message of the StartForkedProcessException content of the temporary
files (together with full jvm command line).