SCHEDULING-256 Concurrent access to the same job log can lead to logs loss
Problem:
When Scheduler.listenJobLogs is called two times for the same job then
second appender doesn't receive log messages produced before
Scheduler.listenJobLogs was called.  

Fix:
First time when SchedulerCore.listenJobLogs is called for some job
SchedulerCore creates AsyncAppender which receives events for running
stored in the TaskLaunchers (it is done in the
TaskLauncher.activateLogs).  When SchedulerCore.listenJobLogs was called
again for the same job new appender was just added to the existing
AsyncAppender and thus didn't receive stored events for running tasks.
To fix it logic of SchedulerCore.listenJobLogs was changed: if then new
client appender is added after AsyncAppender was created then  special
TaskLauncher's method is called to pass stored events to the new client
appender.  

Added new test for this issue.