SCHEDULING-1509 Create a helper application to transfer data at job submission and allow disconnected mode
 Implemented  a VFS file selector (DSFileSelector) which behaves like
dataspace's FastFileSelector .  This selector is used while pushing
input files to the data storage.  An input file will be copied if: -
it  matches one of the includes patterns defined by any of the job's
tasks
	- do not match an exclude pattern
	- if the same pattern is encountered as include and exclude,
the file is copied.

	bug fixed: at initialization, data is pulled only if the job
is finished.

	Functional test for the Proxy has been finished. New classes:

	MyEventListener - listens for events from the Proxy (which
broadcasts events from the Scheduler and adds its own listeners
related to the data transfer)
	TestJavaExecutable - simple JavaExecutable which reads a file from the localspace and creates a
copy to the same location