SCHEDULING-1645 Improved failed selection script handling in the RM
Now if a script throws an exceeption but passes on some nodes
the exception is not propagated to the client. Instead matched nodes
will be returned.

For the scheduler it means job will not become faulty if there is a
node where a selection script generate an exception if there are other
nodes where the script passes.