SCHEDULING-1678 Put tasks and jobs logs in separate files.
In order to display server logs on the client side first
step would be to store logs in separate files for all tasks
jobs.

Storing logs in the data base makes them hard to read on
the server side. That's why file approach was choosen.

The architecture is the following. In scheduler all task
specific logs are written with a special logger that accepts
theis ids and add a perefix "task id" (for job is the same).

This logger is configured to use a regular scheduler appender and
a special appender that appends messages to the file with a name of
task id. It can be desactived from scheduler/rm config file.

Then scheduler passes the task id to the rm (as a computation descriptors)
and RM adds messages to the corresponding file.

In order to simplify the ResourceManager.getNodes call and reduce
the number of parameters a Criteria class was introduced. It is
basically an encapsulation for all the parameters existed before.

In the new FileAppender files are opened/closed for every single
logging event. Probably it should be optimized.

Change-Id: I5eac563f56c50714a99bb5602847f5e65d91d2e1