SCHEDULING-1675 Modify the scripts scheduler-start, rm-start, etc. to do not use eval and use exec instead.
The change was needed by the initd's script to start those script as service.
With eval the script was forking on java start and in that case if you kill
the parent the child was not killed but was detached. With exec the parent
process is replaced by the child process, so no problem to kill the process.

Eval was used for variable expansion, you can do that with exec and arrays
containing the command. This is specific to bash.

We add " around ${JAVACMD[@]} to solve problem with space in path.

Change-Id: If9b719d8d83d8146daa57bb49fa85e39591b699e