SCHEDULING-1692 When job is removed in the scheduler all internal structures must be cleaned.
The scheduler has few lists where it stores its jobs. removeJob request assumes that
user removes only finished job, but not running or pending. So it never cleans the lists
corresponding to such jobs.

The full cleaning was added to removeJob method. If a job is running it will be
terminated first.

Functional test was added verifying the job removal for all job states (well almost for all).

Change-Id: I53d9d1b6ab6cfa45aba641cfdcd52398faef81e2