SCHEDULING-1675 Use "$@" instaead of $@ to properly handle whitespace
From bash manual:

    When the expansion occurs within double quotes, each parameter
    expands to a separate word.  That is, "$@" is equivalent to "$1"
    "$2"

This is exactly what we want. With this change, the escaping required
by rm-client script when creating infrastructures is a bit saner:

Before:

\""\-Dproactive.pamr.router.port=8090 \-Dproactive.scheduler.logs.maxsize=50"\"

After this change:

"\-Dproactive.pamr.router.port=8090 \-Dproactive.scheduler.logs.maxsize=50"

Escaping the first - (dash) is still required because of how options
are parsed in java.

Change-Id: Ib5430b4dc71c53423e606cf967bc6eefa090d024