SCHEDULING-1675 Properly pass VM_ARGS in start-router
The problem was introduced in 8fefec864b9fdd8daf66996021df5881d20049f2:

VM_ARGS was made an array, but it was still referenced as an ordinary
variable when added to JAVACMD. As a result, only the first element of
the array was used.

Fix by reverting to a string. There is no need to use an array for
VM_ARGS since there are no components containing whitespace.

Change-Id: Ib777f51b3ec5a32e5b7a783c9c09c51a9df0b7e6