SCHEDULING-1710  Remove dependency to ProActive.jar in TaskFlowjob API
  - JavaTask.addArgument uses
org.objectweb.proactive.core.util.converteri.ObjectToByteConverter (from
ProActive.jar) in order to convert args to byte
arrays so they will not be deserialized on the Scheduler (which is not
supposed to have the classes corresponding to the args).
The arguments are deserialized in JavaExecutableInitializer#getArguments
by ByteToObjectConverter
The ObjectToByteConverter start a ProActiveRuntime at class load
time.
 We don't need the Programming's ObjectToByteConverter and
ByteToObject converter to do this. Any converter
would do just fine.

 - We created Object2ByteConverter.java to do the 2 wat conversions
between byte[] and Object.

	- in scheduler-build.xml,  add Object2ByteConverter.java
in the filset for target deploy.scheduler.worker

Change-Id: Ib1cbbad298f83f07e73dcdda69309a640199a05f